docs: add Vercel deployment configuration for monorepo
- Add detailed Vercel settings table with all required configuration
- Include step-by-step deployment instructions
- Specify Node.js 20.x as the recommended version
- Document environment variables needed for deployment
- Add notes about standalone output and workspace transpilation
