--- apps/storybook-native/package.json
@@ -1,1 +1,0 @@
-{}
