--- apps/storybook-web/package.json
@@ -1,1 +1,0 @@
-{}
