+++ apps/web/package.json
@@ -2,7 +2,8 @@
   "name": "web",
   "private": true,
   "scripts": {
-    "dev": "next dev -p 3000",
+    "dev": "NODE_OPTIONS='--max-old-space-size=4096' next dev -p 3000 --turbo",
+    "dev:fast": "NODE_OPTIONS='--max-old-space-size=2048' next dev -p 3000 --turbo --no-lint",
     "build": "next build",
     "start": "next start -p 3000",
     "lint": "eslint .",
@@ -10,10 +11,13 @@
     "typecheck": "tsc --noEmit"
   },
   "dependencies": {
-    "next": "15.3.3",
-    "react": "^19.0.0",
-    "react-dom": "^19.0.0",
+    "@emotion/react": "^11.14.0",
+    "@emotion/styled": "^11.14.0",
+    "@mui/icons-material": "^7.1.1",
+    "@mui/material": "^7.1.1",
+    "@mui/material-nextjs": "^7.1.1",
     "@mui/system": "^7.3.1",
+    "@mui/x-date-pickers": "^8.9.0",
     "@smart-desk/data-access": "workspace:*",
     "@smart-desk/design-system": "workspace:*",
     "@smart-desk/hooks": "workspace:*",
@@ -22,6 +26,37 @@
     "@smart-desk/types": "workspace:*",
     "@smart-desk/ui-web": "workspace:*",
     "@smart-desk/utils": "workspace:*",
-    "@supabase/auth-helpers-nextjs": "^0.10.0"
+    "@supabase/auth-helpers-nextjs": "^0.10.0",
+    "@supabase/supabase-js": "^2.50.3",
+    "@tanstack/react-query": "^5.80.7",
+    "axios": "^1.10.0",
+    "cloudinary": "^2.7.0",
+    "date-fns": "^4.1.0",
+    "dompurify": "^3.2.6",
+    "html-react-parser": "^5.2.6",
+    "ical.js": "^2.1.0",
+    "luxon": "^3.6.1",
+    "next": "15.3.3",
+    "next-auth": "^4.24.11",
+    "notistack": "^3.0.2",
+    "openmeteo": "^1.2.0",
+    "react": "^19.0.0",
+    "react-dom": "^19.0.0",
+    "react-hook-form": "^7.61.1",
+    "react-sounds": "^1.0.25",
+    "zod": "^4.0.14",
+    "zustand": "^5.0.5"
+  },
+  "devDependencies": {
+    "@hookform/resolvers": "^5.2.1",
+    "@tanstack/query-async-storage-persister": "^5.83.1",
+    "@tanstack/query-sync-storage-persister": "^5.80.7",
+    "@tanstack/react-query-persist-client": "^5.80.7",
+    "@types/geojson": "^7946.0.16",
+    "@types/luxon": "^3.6.2",
+    "@types/node": "^20",
+    "@types/react": "^19",
+    "@types/react-dom": "^19",
+    "@types/webpack": "^5.28.5"
   }
 }
