+++ pnpm-workspace.yaml
@@ -1,7 +1,6 @@
 packages:
   - apps/*
   - packages/*
-  - tooling/*
 
 ignoredBuiltDependencies:
   - nx
