--- tooling/eslint-config/package.json
@@ -1,42 +1,0 @@
-{
-  "name": "@smart-desk/eslint-config",
-  "version": "0.0.0",
-  "private": true,
-  "type": "module",
-  "main": "index.js",
-  "exports": {
-    ".": "./index.js"
-  },
-  "sideEffects": false,
-  "peerDependencies": {
-    "@next/eslint-plugin-next": ">=15.3.3 <16",
-    "@typescript-eslint/eslint-plugin": "^8.0.0",
-    "@typescript-eslint/parser": "^8.0.0",
-    "eslint": "^9.0.0",
-    "eslint-config-next": ">=15.3.3 <16",
-    "eslint-plugin-import": "^2.31.0",
-    "eslint-plugin-react": "^7.37.0",
-    "eslint-plugin-react-hooks": "^5.0.0",
-    "eslint-plugin-react-native": "^5.0.0",
-    "eslint-plugin-simple-import-sort": "^12.1.0",
-    "typescript": ">=5"
-  },
-  "peerDependenciesMeta": {
-    "eslint-config-next": {
-      "optional": true
-    },
-    "@next/eslint-plugin-next": {
-      "optional": true
-    },
-    "eslint-plugin-react-native": {
-      "optional": true
-    }
-  },
-  "dependencies": {
-    "globals": "^15.11.0",
-    "jsonc-eslint-parser": "^2.4.0"
-  },
-  "devDependencies": {
-    "eslint-plugin-json": "^4.0.1"
-  }
-}
